setting up dev env. created new files for testing
