+++ Dockerfile
@@ -16,7 +16,7 @@ RUN apt-get update && DEBIAN_FRONTEND=noninteractive apt-get install -y \
     rm -rf /var/lib/apt/lists/* && \
     /usr/bin/python3 -m pip install --upgrade pip
 
-WORKDIR notebooks/
+#WORKDIR notebooks/
 
 # Copy the required setup files and install the deepcell-tf dependencies
 COPY setup.py README.md requirements.txt /opt/deepcell-tf/
