+++ deepcell/tmp_funcs.py
@@ -1,0 +1,312 @@
+import os
+import cv2
+import imageio
+
+import numpy as np
+import pandas as pd
+
+from scipy.stats import mode
+
+from skimage.external.tifffile import TiffFile
+from skimage.measure import regionprops_table, regionprops
+
+from sklearn.preprocessing import quantile_transform as qt
+
+"""
+Data Loading Functions
+"""
+
+def get_image(file_name):
+    """Read image from file and returns it as a tensor
+
+    Args:
+        file_name (str): path to image file
+
+    Returns:
+        numpy.array: numpy array of image data
+    """
+    ext = os.path.splitext(file_name.lower())[-1]
+    if ext in ['.tif', '.tiff']:
+        return np.float32(TiffFile(file_name).asarray())
+    return np.float32(imread(file_name))
+
+
+def load_patient_data(path):
+    """Load a single data set
+    Args:
+        path: Location of the dataset
+
+    Returns:
+        np.array: The multiplexed imaging dataset for a specific patient
+    """
+
+    channel_list = []
+
+    for _, _, channels in os.walk(path):
+        # Sort the channels so we know what order they are in
+        channels.sort()
+
+        for channel in channels:
+            # Make sure we don't load the segmentations as features!
+            if 'Segmentation' not in channel:
+                channel_path = os.path.join(path, channel)
+                img = get_image(channel_path)
+                channel_list.append(img)
+    patient_data = np.stack(channel_list, axis=-1)
+
+    return patient_data
+
+def load_mibi_data(path, point_list):
+    """Load mibi dataset
+
+    Args:
+        path: Path containing the mibi dataset
+        point_list: Which patients of the mibi dataset should be loaded
+         * point_list = [2, 5, 8, 9, 21, 22, 24, 26, 34, 37, 38, 41]
+
+    Returns:
+        np.array: The mibi dataset for all of the points
+    """
+
+    # Get the full path for each point
+    path_dict = {}
+    for _, points, _ in os.walk(path):
+        for point in points:
+            patient_path = os.path.join(path, point)
+            path_dict[point] = patient_path
+
+    mibi_data = []
+    
+    for point in point_list:
+
+        point = 'Point{}'.format(point)
+        full_path = path_dict[point]
+        patient_data = load_patient_data(full_path)
+
+        # Make sure dimensions are 2048
+        x_dim, y_dim, _ = patient_data.shape
+        if (x_dim != 2048) or (y_dim != 2048):
+            print(point)
+        else:
+            mibi_data.append(patient_data)
+
+    return np.array(mibi_data)
+
+
+def load_celltypes(path, point_list):
+    """Load celltype dataset
+
+    Args:
+        path: Path containing the cell type dataset
+        point_list: Which patients of the mibi/cell dataset should be loaded
+         * point_list = [2, 5, 8, 9, 21, 22, 24, 26, 34, 37, 38, 41]
+
+    Returns:
+        np.array: The mibi dataset for all of the points
+    """
+    celltype_images = []
+    for point in point_list:
+        filename = "P{}_labeledImage.tiff".format(point)
+        fullpath = os.path.join(path, filename)
+        img = imageio.imread(fullpath)
+        celltype_images.append(img)
+    celltypes = np.stack(celltype_images, axis=0)
+    celltypes = np.expand_dims(celltypes, axis=-1)
+    return celltypes
+
+###########################################################################################
+
+"""
+Data handeling
+"""
+
+def get_cell_df(mibi_data, mibi_labels, mibi_celltypes, markers, marker_idx_dict):
+    """Create a data frame that has all cell information in one place
+
+    Args:
+        mibi_data (np.array):
+        mibi_labels (np.array):
+        mibi_celltypes (np.array):
+        markers (list):
+        marker_idx_dict (dictionary):
+
+    Returns:
+        pd.DataFrame (float): cell by feature matrix with batch info and cell type
+
+    """
+    num_batches = mibi_data.shape[0]
+    num_feats = mibi_data.shape[-1]
+
+    cell_df = pd.DataFrame(0.0, index=range(0), columns=range(num_feats + 2))
+    col_names = np.copy(markers).tolist()
+    col_names.insert(0, 'label')
+    col_names.insert(0, 'batch')
+    cell_df.columns = col_names
+    cell_df['cell_type'] = np.nan
+
+    for batch in range(num_batches):
+        mibi_image = mibi_data[batch]
+        label_image = np.squeeze(mibi_labels[batch])
+        type_image = np.squeeze(mibi_celltypes[batch])
+
+        props = regionprops(label_image)
+
+        num_cells = len(props)
+        num_feats = mibi_data.shape[-1]
+
+        batch_df = pd.DataFrame(0.0, index=range(num_cells), columns=range(num_feats + 2))
+        batch_df.rename(columns={0:'batch', 1:'label'}, inplace = True)
+
+        batch_df['batch'] = batch
+
+        for feature in range(num_feats):
+            marker = marker_idx_dict[feature]
+
+            props_table = regionprops_table(label_image,
+                                            intensity_image=mibi_image[..., feature],
+                                            properties=['mean_intensity', 'label'])
+
+            batch_df['label'] = props_table['label']
+            batch_df[feature + 2] = props_table['mean_intensity']
+            batch_df.rename(columns={feature+2: marker}, inplace=True)
+
+        batch_df['cell_type'] = np.nan
+
+        for prop in props:
+            idx = batch_df[batch_df['label'] == prop.label].index[0]
+
+            coords = prop.coords
+            cts = type_image[coords[:, 0], coords[:, 1]]
+            cell_type = mode(cts, axis=None).mode[0]
+
+            batch_df.loc[idx, 'cell_type'] = cell_type
+
+        cell_df = pd.concat([cell_df, batch_df])
+
+    return cell_df
+
+
+def qt_transform(X):
+    """Perform sklearn.preprocessing.quantile_transform on batch data
+
+    Args:
+        X: data to transform
+
+    Returns:
+        np.array: The mibi dataset transformed
+    """
+
+    batches = X.shape[-1]
+    transformed_data = []
+
+    for batch in range(batches):
+        x_batch = X[..., batch]
+
+        x_batch = qt(x_batch,
+                     copy=False,
+                     output_distribution='uniform',
+                     n_quantiles=10)
+
+        transformed_data.append(x_batch)
+
+    return np.array(transformed_data)
+
+
+def normalize_mibi_data(tmp_mibi_data, marker_idx_dict):
+    """Normalize mibi data by applying a gaussian smothing on the raw data
+       removing bottom 5% of labels and breaking remaining values into quantiles
+
+    Args:
+        tmp_mibi_data: raw mibi data loaded via load_mibi func
+        marker_idx_dict: look up table to go from index to maker name
+
+    Returns:
+        np.array: normalized mibi_data
+    """
+
+    num_batches = tmp_mibi_data.shape[0]
+    num_channels = tmp_mibi_data.shape[-1]
+
+    mibi_data = []
+
+    for batch in range(num_batches):
+        mibi_batch = tmp_mibi_data[batch, ...]
+        # channel_data = []
+
+        channel_imgs = []
+        channel_th = []
+        for channel in range(num_channels):
+
+            batch_channel_data = mibi_batch[..., channel]
+            blur_img = cv2.GaussianBlur(batch_channel_data, (3, 3), 0)
+
+            channel_imgs.append(blur_img)
+
+            if len(blur_img[blur_img > 0]) == 0:
+#                 print('batch {}'.format(batch), marker_idx_dict[channel])
+                channel_th.append(0)
+            else:
+                low_vals = np.percentile(blur_img[blur_img > 0], 5)
+                channel_th.append(low_vals)
+
+        imgs = np.array(channel_imgs).T
+
+        imgs.T[np.tile((np.sum((imgs < channel_th).T, axis=0) == num_channels),
+                       (num_channels, 1)).reshape(imgs.T.shape)] = 0
+
+        batch_data = qt_transform(imgs)
+
+        mibi_data.append(batch_data.T)
+
+    return np.array(mibi_data)
+
+
+###########################################################################################
+
+"""
+Auxiliary Functions
+"""
+
+def get_marker_dict(path):
+    """Load a single dataset
+
+    Args:
+        path: Location of the dataset
+
+    Returns:
+        np.array: The multiplexed imaging dataset for a specific patient
+    """
+
+    marker_list = []
+
+    for _, _, channels in os.walk(path):
+        # Sort the channels so we know what order they are in
+        channels.sort()
+        for channel in channels:
+            # Make sure we don't load the segmentations as features!
+            if 'Segmentation' not in channel:
+                marker_list.append(channel.split('.')[0])
+
+    marker_dict = dict(zip(marker_list, list(range(len(marker_list)))))
+    # marker_by_idx_dict = dict(zip(list(range(len(marker_list))),marker_list))
+
+    return marker_dict
+
+
+
+# def celltype_to_labeled_img(mibi_celltypes, celltype_data):
+#     """
+#     DOC STRING
+#     """
+
+#     new_label_image = np.zeros(mibi_celltypes.shape)
+#     for batch in range(mibi_celltypes.shape[0]):
+#         print(batch)
+#         props = regionprops(mibi_celltypes[batch, ..., 0])
+#         for i, prop in enumerate(props):
+#             nli_batch = new_label_image[batch]
+#             nli_batch[prop.coords[:, 0], prop.coords[:, 1]] = celltype_data[batch, i]
+#         new_label_image[batch] = nli_batch
+
+
+#     return new_label_image
