+++ requirements.txt
@@ -12,4 +12,4 @@ pydot>=1.4.1
 pydotplus>=2.0.2
 graphviz>=0.14
 einops==0.3.0
-
+jupyterlab
